Merged PR 263: Azure Functions Optimisers
This is available at dev -  https://app73204e59.azurewebsites.net/

When creating an Item Recommender, you can select "auto-AI", which connects the recommender to a backend Azure Functions instance containing "optimisers".

Optimisers are stored in Tables in Azure Storage.
Optimiser parameters are stored in Azure Blobs.
The functions are bound to the table row and azure blob via the path.
Optimisers are tenant aware.

For now

- add script to run azureite container
- move to import "cherry.ai"
- enable use auto AI
- update UI for auto-ai
- add connection to python functions
- improve create and invoke for #4578
- update funloader with cherry colour
- dont print so much
- update status box w/ auto ai
- fix: no synapse broke the kv access policies
- remove unused config

Related work items: #4578