Merged PR 364: Add attach to chrome config
- Add attach to chrome config

To debug, we need to run .NET Core Launch

![image.png](https://dev.azure.com/cherry-ai/fc25747b-9870-49d7-b9a7-b98a1e1e439d/_apis/git/repositories/e336e387-c1f5-4246-a6c1-e9d790719921/pullRequests/364/attachments/image.png)

Breakpoints to the web app should be possible but not on the client app.

We need to run Attach to Chrome while the debug is on going and it will launch a new chrome browser in which breakpoints on the client app should be possible.

![image (2).png](https://dev.azure.com/cherry-ai/fc25747b-9870-49d7-b9a7-b98a1e1e439d/_apis/git/repositories/e336e387-c1f5-4246-a6c1-e9d790719921/pullRequests/364/attachments/image%20%282%29.png)

Related work items: #5298