Merged PR 571: Connect to Klaviyo
Klaviyo integration requires API keys to communicate with. The API keys will either be part of the payload or passed in the query parameter of the endpoint when sending requests.

Public api key is used to 'Track and Identify' APIs and this does not adhere to REST principles. Private key is used to manage data (Profiles, Metrics, etc) within the Klaviyo account and adheres to REST principles.

Sample endpoint for getting list:  _'https://a.klaviyo.com/api/v2/lists?api_key=API_KEY'_

- Added Klaviyo as Integrated System
- Added page to configure Klaviyo API keys
- API Keys can be set multiple times but values will not be shown in the textboxes
- IntegrationStatus.OK will be set once API keys are saved

**Create Integration**

![Screen Shot 2022-04-25 at 3.02.23 PM.png](https://dev.azure.com/cherry-ai/fc25747b-9870-49d7-b9a7-b98a1e1e439d/_apis/git/repositories/e336e387-c1f5-4246-a6c1-e9d790719921/pullRequests/571/attachments/Screen%20Shot%202022-04-25%20at%203.02.23%20PM.png) /Users/cherry.gonzales/Desktop/Screen Shot 2022-04-25 at 3.02.23 PM.png

**Klaviyo Integrated System Details**

![image (3).png](https://dev.azure.com/cherry-ai/fc25747b-9870-49d7-b9a7-b98a1e1e439d/_apis/git/repositories/e336e387-c1f5-4246-a6c1-e9d790719921/pullRequests/571/attachments/image%20%283%29.png)

![image (4).png](https://dev.azure.com/cherry-ai/fc25747b-9870-49d7-b9a7-b98a1e1e439d/_apis/git/repositories/e336e387-c1f5-4246-a6c1-e9d790719921/pullRequests/571/attachments/image%20%284%29.png)

Related work items: #5560