Merged PR 575: Additional changes to Shopify billing
- Webhook sets Shopify integration status to OK
- Fixed UserError Field property type
- Don't force on multitenant
- Use navigate instead of window.location.href

For the scenario where the user neither accepts or declines a recurring charge, we moved the integration status logic to the Shopify webhook workflows.
(Downside to this approach is we expect the webhook to run fast enough before the user navigates to the integration detail page. Since a login prompt is required, we don't expect this to be an issue)

When in multi-tenant mode, there is an issue of being unable to determine if the environments have loaded correctly. When the tenant has multiple environments, the installation was being forced unto the default environment. To resolve, we remove the auto connect for multi-tenant mode.

Changed `window.location.href` to `navigate` to prevent the extra login prompt. Tested on both multi-tenant and single tenant mode.

Related work items: #5670